Update script.js
Found bug that at the end of the trip the calculations were not correct and added one day to the trip both with "T" and "O". Effectivly adding to the Off at home time. and duplicating the enddate twice as the same day, but one with a T and one with O. 
fixed.